feat: change response into response entity

+++ src/main/java/com/example/comitserver/controller/StudyController.java
@@ -1,11 +1,13 @@
 package com.example.comitserver.controller;
 
 import com.example.comitserver.dto.CustomUserDetails;
+import com.example.comitserver.dto.ServerResponseDTO;
 import com.example.comitserver.dto.StudyRequestDTO;
 import com.example.comitserver.dto.StudyResponseDTO;
 import com.example.comitserver.entity.StudyEntity;
 import com.example.comitserver.repository.StudyRepository;
 import com.example.comitserver.service.StudyService;
+import com.example.comitserver.utils.ResponseUtil;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -24,30 +26,36 @@ public class StudyController {
 
     private final StudyService studyService;
     private final ModelMapper modelMapper;
+    private final StudyRepository studyRepository;
 
     @Autowired
-    public StudyController(StudyService studyService, ModelMapper modelMapper) {
+    public StudyController(StudyService studyService, ModelMapper modelMapper, StudyRepository studyRepository) {
         this.studyService = studyService;
         this.modelMapper = modelMapper;
+        this.studyRepository = studyRepository;
     }
 
     @GetMapping("/studies")
-    public ResponseEntity<List<StudyResponseDTO>> getStudies() {
+    public ResponseEntity<ServerResponseDTO> getStudies() {
         List<StudyEntity> allStudies = studyService.showAllStudies();
         List<StudyResponseDTO> allStudiesDTO = allStudies.stream().map(entity -> modelMapper.map(entity, StudyResponseDTO.class)).collect(Collectors.toList());
 
-        return ResponseEntity.ok(allStudiesDTO);
+        return ResponseUtil.createSuccessResponse(allStudiesDTO, HttpStatus.OK);
+        //return ResponseEntity.ok(allStudiesDTO);
     }
 
     @GetMapping("/studies/{id}")
-    public ResponseEntity<StudyResponseDTO> getStudyById(@PathVariable Long id) {
+    public ResponseEntity<ServerResponseDTO> getStudyById(@PathVariable Long id) {
+        if (studyRepository.findById(id).isEmpty()) return ResponseUtil.createErrorResponse(HttpStatus.NOT_FOUND, "Study/CannotFindId", "study with that id not found");
+
         StudyEntity study = studyService.showStudy(id);
 
-        return ResponseEntity.ok(modelMapper.map(study, StudyResponseDTO.class));
+        return ResponseUtil.createSuccessResponse(study, HttpStatus.OK);
+        //return ResponseEntity.ok(modelMapper.map(study, StudyResponseDTO.class));
     }
 
     @PostMapping("/studies")
-    public ResponseEntity<StudyResponseDTO> postStudy(@RequestBody StudyRequestDTO studyRequestDTO, @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+    public ResponseEntity<ServerResponseDTO> postStudy(@RequestBody StudyRequestDTO studyRequestDTO, @AuthenticationPrincipal CustomUserDetails customUserDetails) {
         StudyEntity newStudy = studyService.createStudy(studyRequestDTO, customUserDetails);
 
         URI location = ServletUriComponentsBuilder
@@ -56,34 +64,44 @@ public class StudyController {
                 .buildAndExpand(newStudy.getId())
                 .toUri();
 
-        return ResponseEntity.created(location).body(modelMapper.map(newStudy, StudyResponseDTO.class));
+        return ResponseUtil.createSuccessResponse(modelMapper.map(newStudy, StudyResponseDTO.class), HttpStatus.CREATED, location);
+        //return ResponseEntity.created(location).body(modelMapper.map(newStudy, StudyResponseDTO.class));
     }
 
     @PutMapping("/studies/{id}")
-    public ResponseEntity<StudyResponseDTO> putStudy(@PathVariable Long id, @RequestBody StudyRequestDTO studyRequestDTO, @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+    public ResponseEntity<ServerResponseDTO> putStudy(@PathVariable Long id, @RequestBody StudyRequestDTO studyRequestDTO, @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        if (studyRepository.findById(id).isEmpty()) return ResponseUtil.createErrorResponse(HttpStatus.NOT_FOUND, "Study/CannotFindId", "study with that id not found");
+
         if (studyService.identification(id, customUserDetails)) {
             StudyEntity updatedStudy = studyService.updateStudy(id, studyRequestDTO);
 
             if (updatedStudy == null) {
-                return ResponseEntity.notFound().build();
+                return ResponseUtil.createErrorResponse(HttpStatus.NOT_FOUND, "Study/CannotFindId", "the result of updating study is null");
+                //return ResponseEntity.notFound().build();
             }
 
-            return ResponseEntity.ok(modelMapper.map(updatedStudy, StudyResponseDTO.class));
+            return ResponseUtil.createSuccessResponse(modelMapper.map(updatedStudy, StudyResponseDTO.class), HttpStatus.OK);
+            //return ResponseEntity.ok(modelMapper.map(updatedStudy, StudyResponseDTO.class));
         }
-        else return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        else return ResponseUtil.createErrorResponse(HttpStatus.FORBIDDEN, "Study/PermissionDenied", "the user does not have permission to update this study");
+                //ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
     @DeleteMapping("/studies/{id}")
-    public ResponseEntity<StudyResponseDTO> deleteStudy(@PathVariable Long id, @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+    public ResponseEntity<ServerResponseDTO> deleteStudy(@PathVariable Long id, @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        if (studyRepository.findById(id).isEmpty()) return ResponseUtil.createErrorResponse(HttpStatus.NOT_FOUND, "Study/CannotFindId", "study with that id not found");
+
         if(studyService.identification(id, customUserDetails)) {
             StudyEntity deletedStudy = studyService.showStudy(id);
             StudyResponseDTO studyResponseDTO = modelMapper.map(deletedStudy, StudyResponseDTO.class);
 
             studyService.deleteStudy(id);
 
-            return ResponseEntity.ok(studyResponseDTO);
+            return ResponseUtil.createSuccessResponse(studyResponseDTO, HttpStatus.OK);
+            //return ResponseEntity.ok(studyResponseDTO);
         }
-        else return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        else return ResponseUtil.createErrorResponse(HttpStatus.FORBIDDEN, "Study/PermissionDenied", "the user does not have permission to delete this study");
+                //ResponseEntity.status(HttpStatus.FORBIDDEN).build();
     }
 
 }
